:memo: :: fix total count

+++ src/main/java/dgsw/pioneers/checkIn/domain/suggestion/adapter/in/web/dto/res/SuggestionWithTotalCountRes.java
@@ -10,6 +10,6 @@ import java.util.List;
 @Builder
 public class SuggestionWithTotalCountRes {
 
-    private int totalCount;
+    private Long totalCount;
     private List<Suggestion> value;
 }
